% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate report}
\usage{
generate_report(
  unique_id = NULL,
  credentials_file = NULL,
  date = NULL,
  output_dir = NULL,
  output_file = NULL,
  which_days = NULL
)
}
\arguments{
\item{unique_id}{The unique ID of the case}

\item{credentials_file}{Path to credentials file}

\item{date}{A date to be printed on the report. If \code{NULL} (the
default), the current date will be used}

\item{output_dir}{The directory to which the file should be written. If
\code{NULL} (the default), the current working directory will be used.}

\item{output_file}{The name of the file to be written.}

\item{which_days}{A date of a vector of dates to view on the report}
}
\value{
An html will be written
}
\description{
Generate report following a positive case
}
