% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cases.R
\name{get_cases}
\alias{get_cases}
\title{Get cases}
\usage{
get_cases(dbname, host, user, pass, port = 3306, unique_id = NULL)
}
\arguments{
\item{dbname}{The name of your database}

\item{host}{The host ip address of your database}

\item{user}{The username for your database}

\item{pass}{The password for your database}

\item{port}{The port number for yourdatabse}

\item{unique_id}{Optional, restrict to only one unique_id}
}
\value{
A dataframe
}
\description{
Retrieve the table of reported cases
This is NOT part of the traccar suite of tools
}
