% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_positions_from_unique_id.R,
%   R/get_reports_from_unique_id.R
\name{get_device_id_from_unique_id}
\alias{get_device_id_from_unique_id}
\title{Get device id from unique id}
\usage{
get_device_id_from_unique_id(url, user, pass, unique_id, path = "api/devices")

get_device_id_from_unique_id(url, user, pass, unique_id, path = "api/devices")
}
\arguments{
\item{url}{The url of your traccar api}

\item{user}{The username for your traccar api}

\item{pass}{The password for your traccar api}

\item{unique_id}{The unique id for a regestered HCW on the traccar api}

\item{path}{API path}
}
\value{
A dataframe

A dataframe
}
\description{
Retrieve device id from api/devices using unique id

Retrieve device id from api/devices using unique id
}
