% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_registered_workers.R
\name{get_registered_workers}
\alias{get_registered_workers}
\title{Get registered workers}
\usage{
get_registered_workers(dbname, host, user, pass, port = 3306)
}
\arguments{
\item{dbname}{The name of your database}

\item{host}{The host ip address of your database}

\item{user}{The username for your database}

\item{pass}{The password for your database}

\item{port}{The port number for yourdatabse}
}
\value{
A dataframe
}
\description{
Retrieve registered workers from the database associated wtih the shiny app
This is NOT part of the traccar suite of tools
}
