% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_database.R
\name{set_up_database}
\alias{set_up_database}
\title{Set up the database}
\usage{
set_up_database(credentials)
}
\arguments{
\item{credentials}{The path to a credentials file}
}
\value{
Database tables will be set up
}
\description{
Set up the databse. Only needs to be run one
}
