% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_workers.R
\name{sync_workers}
\alias{sync_workers}
\title{Sync workers}
\usage{
sync_workers(
  credentials = NULL,
  delete_from_traccar = FALSE,
  dbname,
  host,
  user,
  pass,
  port,
  traccar_url,
  traccar_user,
  traccar_pass
)
}
\arguments{
\item{credentials}{A list with the following named elements: dbname, host, user, pass, port (all referring to the MySQL database which communicates with the shiny app) and traccar_url, traccar_user, traccar_pass (these last 3 referring to the Traccar server). If \code{NULL} these arguments can be specified individually}

\item{delete_from_traccar}{Whether to delete from Traccar those workers who don't show up in the shiny database. Be careful with this!}

\item{dbname}{Name of the MySQL database which communicates with the Shiny server. Example: bcvdb}

\item{host}{The host (IP) of the MySQL database which communicates with the Shiny server}

\item{user}{The username of the MySQL database which communicates wih the Shiny server}

\item{pass}{The password for accessing the MySQL database which communicates wih the Shiny server}

\item{port}{The port of the MySQL database which communicates with the Shiny Server}

\item{traccar_url}{The URL where the Traccar server application is being run}

\item{traccar_user}{The administrator's username for the Traccar server application}

\item{traccar_pass}{The administrator's password for the Traccar server application}
}
\value{
The function will send updates from the Shiny-generated MySQL database to the Traccar server via an HTTP post request
}
\description{
Syncronize between (a) the workers who have registered via the shiny app and (b) the workers whose information has been registered on the Traccar server app
}
\examples{
\dontrun{
# Read in credentials
credentials <- yaml::yaml.load_file('../credentials/credentials.yaml')
sync_workers(credentials = credentials)
}
}
